fixup: kconfiglib exception on vscode terminal
For the vscode terminal on remote server,
there is a chance that the client language
is not identical to server. And hence the
language required may not be supported
by server environment. Finally,
we will see an exception occurred with
kconfiglib.

To fix this issue, this patch will check if
the LANG is supported. And set a fall
back option if not.

See https://github.com/RT-Thread/rt-thread/pull/8843.

Signed-off-by: Shell <smokewood@qq.com>